update 20/07/03 오후 1:57분

+++ byeolsolResort/src/controller/CustomerController.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.sql.Date;
 import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.Enumeration;
 import java.util.List;
 
@@ -14,7 +15,9 @@ import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import javax.validation.constraints.Pattern;
 
+import org.quartz.impl.calendar.HolidayCalendar;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.format.datetime.DateFormatter;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Controller;
@@ -36,11 +39,16 @@ import model.dto.Customer;
 import model.dto.CustomerVO;
 import model.service.BoardService;
 import model.service.CustomerService;
+import model.service.HolyDayService;
 import model.service.ReservService;
 
 @Controller
 @RequestMapping("/cus")
 public class CustomerController {
+
+	private static final String DEFAULT_VALUE_ID = "defaultValueId";
+	private static final String DEFAULT_VALUE_PASSWORD = "defaultValuePassword";
+	private static final String DEFAULT_VALUE_Email = "defaultValueEmail@test.com";
 
 	@Autowired
 	CustomerService customerService;
@@ -90,8 +98,7 @@ public class CustomerController {
 						String emailCheck = customerService.emailCheck(customerVo.getEmail());
 						if (!emailCheck.equals("중복")) {
 							if (!customerService.phoneCheck(customerVo.getPhone()).equals("중복")) {
-								if (LocalDate.now().getYear()-birth.toLocalDate().getYear()>=5) {
-									System.out.println("맞음");
+								if (LocalDate.now().getYear() - birth.toLocalDate().getYear() >= 5) {
 									Customer customer = new Customer(customerVo.getId(), customerVo.getUserId(),
 											customerVo.getPassword(), customerVo.getName(), customerVo.getZipCode(),
 											customerVo.getEmail(), customerVo.getAddress(),
@@ -115,8 +122,8 @@ public class CustomerController {
 						}
 					}
 				}
-			}else {
-				m.addAttribute("errorMessage","생년월일을 다시 확인 해주세요");
+			} else {
+				m.addAttribute("errorMessage", "생년월일을 다시 확인 해주세요");
 				return "redirect:/cus/regis";
 			}
 
@@ -158,18 +165,24 @@ public class CustomerController {
 	}
 
 	@PostMapping("/login")
-	public String login(String userId, String password, HttpSession session, Model m) {
-		Customer c = customerService.logIn(userId, password);
-		if (c == null) {
-			m.addAttribute("islog", "id 또는 비밀번호가 잘못 되었습니다.");
-			return "loginForm";
-		} else {
+	public String login(@RequestParam(defaultValue = DEFAULT_VALUE_ID) String userId,
+			@RequestParam(defaultValue = DEFAULT_VALUE_PASSWORD) String password, HttpSession session, Model m) {
+		if (!userId.equals(DEFAULT_VALUE_ID) || !password.equals(DEFAULT_VALUE_PASSWORD)) {
+			Customer c = customerService.logIn(userId, password);
+			if (c == null) {
+				m.addAttribute("islog", "id 또는 비밀번호가 잘못 되었습니다.");
+				return "loginForm";
+			} else {
 
-			session.setAttribute("userId", c.getUserId());
-			session.setAttribute("userName", c.getName());
-			session.setAttribute("state", c.getEmailState());
-			session.setAttribute("userEmail", c.getEmail());
-			return "redirect:/index/main";
+				session.setAttribute("userId", c.getUserId());
+				session.setAttribute("userName", c.getName());
+				session.setAttribute("state", c.getEmailState());
+				session.setAttribute("userEmail", c.getEmail());
+				return "redirect:/index/main";
+			}
+		} else {
+			m.addAttribute("errorMessage", "id 또는 비밀번호를 확인해 주세요");
+			return "redirect:/cus/login";
 		}
 	}
 
@@ -203,12 +216,12 @@ public class CustomerController {
 
 	@PostMapping("/updateCustomer")
 	public String updateCustomer(HttpSession session, Model m,
-			@ModelAttribute("CustomerVo") @Valid CustomerVO customerVo, BindingResult result, Date birth) {
-		System.out.println(birth);
+			@ModelAttribute("CustomerVo") @Valid CustomerVO customerVo, BindingResult result, @RequestParam(defaultValue = DEFAULT_BIRTH_DATE)Date birth) {
 		if (session.getAttribute("userId") != null) {
 			String userId = (String) session.getAttribute("userId");
 			Customer cust = new Customer();
 			if (customerVo.getUserId().equals(userId)) {
+				if(!birth.toString().equals(DEFAULT_BIRTH_DATE)) {
 				Customer customer = customerService.getCustomerById(userId);
 				cust.setId(customer.getId());
 				cust.setUserId(userId);
@@ -227,7 +240,7 @@ public class CustomerController {
 					} else {
 						cust.setPhone(customerVo.getPhone());
 					}
-					if (LocalDate.now().getYear()-birth.toLocalDate().getYear()>=5) {
+					if (LocalDate.now().getYear() - birth.toLocalDate().getYear() >= 5) {
 						cust.setBirthDate(birth.toLocalDate());
 					}
 					cust.setAddress(customerVo.getAddress());
@@ -243,7 +256,10 @@ public class CustomerController {
 					customerService.updateCustomer(cust);
 					return "redirect:/cus/myPage";
 				}
-
+				}else {
+					m.addAttribute("errorMessage","생일을 다시 확인 하여 주세요");
+					return "redirect:/cus/myPage";
+				}
 			} else {
 				m.addAttribute("errorMessage", "권한이 없습니다.");
 				return "redirect:/index/main";
@@ -272,9 +288,13 @@ public class CustomerController {
 
 	@PostMapping(value = "/mailSend", produces = "application/text;charset=utf-8")
 	@ResponseBody
-	public String mailSendRes(String userEmail, HttpSession session) {
+	public String mailSendRes(@RequestParam(defaultValue = DEFAULT_VALUE_Email)String userEmail, HttpSession session) {
+		if(!userEmail.equals(DEFAULT_VALUE_Email)) {
 		customerService.mailSend(mailSender, userEmail);
 		return "이메일 인증코드를 성공적으로 보냈습니다.";
+		}else {
+			return "잘못된 접근";
+		}
 	}
 
 	@PostMapping("/mailCheck")
@@ -296,7 +316,7 @@ public class CustomerController {
 	ReservService reservService;
 
 	@GetMapping("/myReserv")
-	public String myReservPage(HttpSession session, int pageNum, Model m) {
+	public String myReservPage(HttpSession session, @RequestParam(defaultValue = "1") int pageNum, Model m) {
 		if (session.getAttribute("userId") != null) {
 			String userId = (String) session.getAttribute("userId");
 			Customer customer = customerService.getCustomerById(userId);
@@ -312,42 +332,47 @@ public class CustomerController {
 			return "redirect:/index/main";
 		}
 	}
-	
+
 	@GetMapping("/deleteCustomer")
 	public String deleteCustomerForm(HttpSession session, Model m) {
-		if(session.getAttribute("userId")!=null) {
-			return "deleteForm";
-		}else {
-			m.addAttribute("errorMessage","로그인이 되어 있지 않습니다.");
+		if (session.getAttribute("userId") != null) {
+			return "customerDeleteForm";
+		} else {
+			m.addAttribute("errorMessage", "로그인이 되어 있지 않습니다.");
 			return "redirect:/index/main";
 		}
 	}
-	
-	
+
 	@PostMapping("/deleteCustomer")
-	public String deleteCustomer(HttpSession session, String userId, String password,Model m) {
-		if(session.getAttribute("userId")!=null) {
-			String sessionUserId = (String)session.getAttribute("userId");
-			Customer sessionCustoemr = customerService.getCustomerById(sessionUserId);
-			if(sessionCustoemr.getPassword().equals(password)) {
-				customerService.deleteCustomerWithAllInfor(sessionCustoemr);
-				session.setAttribute("userId", null);
-				session.setAttribute("userName", null);
-				session.setAttribute("state", null);
-				session.setAttribute("userEmail", null);
-				m.addAttribute("deleteMessage","계정을 탈퇴 하였습니다");
-				return "redirect:/index/main";
-			}else {
-				m.addAttribute("errorMessage","비밀번호를 확인해 주세요");
+	public String deleteCustomer(HttpSession session, @RequestParam(defaultValue = DEFAULT_VALUE_ID) String userId,
+			@RequestParam(defaultValue = DEFAULT_VALUE_PASSWORD) String password, Model m) {
+		if (session.getAttribute("userId") != null) {
+			if (!userId.equals(DEFAULT_VALUE_ID) || !password.equals(DEFAULT_VALUE_PASSWORD)) {
+				String sessionUserId = (String) session.getAttribute("userId");
+				Customer sessionCustoemr = customerService.getCustomerById(sessionUserId);
+				if (sessionCustoemr.getPassword().equals(password)) {
+					customerService.deleteCustomerWithAllInfor(sessionCustoemr);
+					session.setAttribute("userId", null);
+					session.setAttribute("userName", null);
+					session.setAttribute("state", null);
+					session.setAttribute("userEmail", null);
+					m.addAttribute("deleteMessage", "계정을 탈퇴 하였습니다");
+					return "redirect:/index/main";
+				} else {
+					m.addAttribute("errorMessage", "아이디 또는 비밀번호를 다시 확인 해 주세요");
+					return "redirect:/cus/deleteCustomer";
+				}
+			} else {
+				m.addAttribute("errorMessage", "비밀번호를 확인해 주세요");
 				return "redirect:/cus/deleteCustomer";
 			}
-		}else {
-			m.addAttribute("errorMessage","로그인이 되어 있지 않습니다.");
+		} else {
+			m.addAttribute("errorMessage", "로그인이 되어 있지 않습니다.");
 			return "redirect:/index/main";
 		}
-		
-		
+
 	}
+	
 	
 
 }
